+++ tests/fixtures/api/auth/login/_.POST.json
@@ -1,1 +1,3 @@
-{}
+{
+    "token": "my_token"
+}
